ENH: Make homogeneous the usage of degrees and radians
The best solution for the moment that I have found is to only convert from deg to rad just before calling a trigonometric method (or approximation, like in the lift coefficient).

Like that, it becomes easy to debug (we are used to read degrees, not radians) and the code is "robust'. Ideally there should be a class that defines `Angle` and computes correctly the trigonometric functions.